+++ packages/falcon-ecommerce-uikit/package.json
@@ -41,7 +41,7 @@
     "qs": "6.6.0",
     "react-adopt": "^0.6.0",
     "react-credit-card-input": "1.1.1",
-    "react-lazyload": "2.5.0",
+    "react-lazyload": "2.6.2",
     "react-powerplug": "1.0.0"
   },
   "devDependencies": {
@@ -56,7 +56,7 @@
     "@types/lodash.isequal": "^4.5.3",
     "@types/react": "16.8.22",
     "@types/react-dom": "16.8.4",
-    "@types/react-lazyload": "2.3.2",
+    "@types/react-lazyload": "2.5.0",
     "@types/react-router-dom": "4.3.4",
     "@types/react-test-renderer": "16.8.2",
     "apollo-cache-inmemory": "1.5.1",
