+++ packages/falcon-ecommerce-uikit/src/ProductList/index.tsx
@@ -1,3 +1,1 @@
 export * from './ProductList';
-export * from './ProductCard';
-export * from './EmptyProductList';
