+++ packages/falcon-ui-kit/package.json
@@ -29,6 +29,7 @@
     "@deity/falcon-shop-data": "^0.0.1",
     "core-js": "3.1.4",
     "formik": "^1.5.8",
+    "react-lazyload": "2.6.2",
     "react-powerplug": "1.0.0",
     "react-responsive": "6.1.1"
   },
@@ -39,6 +40,7 @@
     "@deity/falcon-front-kit": "^0.0.1",
     "@types/react": "16.8.22",
     "@types/react-dom": "16.8.4",
+    "@types/react-lazyload": "2.5.0",
     "@types/react-responsive": "3.0.3",
     "@types/react-router-dom": "4.3.4",
     "graphql-tag": "2.10.1",
